+++ week03/1-to-50/src/App.jsx
@@ -3,8 +3,8 @@ import { ThemeProvider } from '@emotion/react';
 import theme from './styles/theme';
 import { GlobalStyles } from './styles/GlobalStyle';
 import Header from './components/Header';
-import GameBoard from './components/GameBoard';
-import RankingBoard from './components/RankingBoard';
+import GameBoard from './components/CardGame/GameBoard';
+import RankingBoard from './components/Ranking/RankingBoard';
 import Modal from './components/Modal';
 import { useTimer } from './utils/timer';
 import { saveGameData, loadGameData } from './utils/storage';
