+++ week03/1-to-50/src/utils/cardUtil.js
@@ -1,5 +1,3 @@
-// cardUtil.js
-
 export const generateShuffledCards = (firstSet, secondSet) => {
   const firstNumbers = Array.from(
     { length: firstSet },
